Fix performance

+++ nacossync-worker/src/main/java/com/alibaba/nacossync/extension/impl/NacosSyncToZookeeperServiceImpl.java
@@ -30,13 +30,10 @@ import com.alibaba.nacossync.extension.impl.extend.Sharding;
 import com.alibaba.nacossync.monitor.MetricsManager;
 import com.alibaba.nacossync.pojo.model.TaskDO;
 import com.alibaba.nacossync.util.DubboConstants;
-import com.alibaba.nacossync.util.ExpirySet;
 import com.google.common.collect.Sets;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.recipes.cache.PathChildrenCache;
-import org.apache.curator.framework.recipes.cache.PathChildrenCacheEvent;
-import org.apache.curator.utils.CloseableUtils;
 import org.apache.zookeeper.CreateMode;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -47,6 +44,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import static com.alibaba.nacossync.constant.SkyWalkerConstants.SOURCE_CLUSTERID_KEY;
 import static com.alibaba.nacossync.util.StringUtils.convertDubboFullPathForZk;
 import static com.alibaba.nacossync.util.StringUtils.convertDubboProvidersPath;
 
@@ -93,9 +91,7 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
 
     private final ZookeeperServerHolder zookeeperServerHolder;
 
-    private static ExpirySet<String> serviceNameSet = new ExpirySet<String>();
-
-    private static ExecutorService EXECUTOR = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
+    private static ExecutorService EXECUTOR = Executors.newFixedThreadPool(100);
 
     @Resource(type = NacosSyncToZookeeperServicesSharding.class)
     private Sharding sharding;
@@ -112,14 +108,14 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
     public boolean delete(TaskDO taskDO) {
         try {
             NamingService sourceNamingService =
-                    //nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getGroupName());//
                     nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getNameSpace());//fix with no nameSpaceName
             EventListener eventListener = nacosListenerMap.remove(taskDO.getTaskId());
-            PathChildrenCache pathChildrenCache = pathChildrenCacheMap.get(taskDO.getTaskId());
             sourceNamingService.unsubscribe(taskDO.getServiceName(), eventListener);
-            CloseableUtils.closeQuietly(pathChildrenCache);
             Set<String> instanceUrlSet = instanceBackupMap.get(taskDO.getTaskId());
             CuratorFramework client = zookeeperServerHolder.get(taskDO.getDestClusterId(), taskDO.getGroupName());
+            if (CollectionUtils.isEmpty(instanceUrlSet)) {
+                return true;
+            }
             for (String instanceUrl : instanceUrlSet) {
                 client.delete().quietly().forPath(instanceUrl);
             }
@@ -136,30 +132,6 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
     public boolean sync(TaskDO taskDO) {
         sharding.start(taskDO);
         return true;
-    }
-
-    private void tryToCompensate(TaskDO taskDO, NamingService sourceNamingService, List<Instance> sourceInstances) {
-        if (!CollectionUtils.isEmpty(sourceInstances)) {
-            final PathChildrenCache pathCache = getPathCache(taskDO);
-            if (pathCache.getListenable().size() == 0) { // 防止重复注册
-                pathCache.getListenable().addListener((zkClient, zkEvent) -> {
-                    if (zkEvent.getType() == PathChildrenCacheEvent.Type.CHILD_REMOVED) {
-                        List<Instance> allInstances =
-                                sourceNamingService.getAllInstances(taskDO.getServiceName());
-                        for (Instance instance : allInstances) {
-                            String instanceUrl = buildSyncInstance(instance, taskDO);
-                            String zkInstancePath = zkEvent.getData().getPath();
-                            if (zkInstancePath.equals(instanceUrl)) {
-                                zkClient.create().creatingParentsIfNeeded().withMode(CreateMode.EPHEMERAL)
-                                        .forPath(zkInstancePath);
-                                break;
-                            }
-                        }
-                    }
-                });
-            }
-
-        }
     }
 
     private void deleteInvalidInstances(TaskDO taskDO, CuratorFramework client, Set<String> newInstanceUrlSet)
@@ -179,12 +151,25 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
         HashSet<String> waitingToAddInstance = new HashSet<>();
         for (Instance instance : sourceInstances) {
             if (needSync(instance.getMetadata())) {
-                log.info("nacos->zk ,real sync service :{},and instance :{}", instance.getServiceName(), instance.getIp());
                 String instanceUrl = buildSyncInstance(instance, taskDO);
-                if (null == client.checkExists().forPath(instanceUrl)) {
-                    client.create().creatingParentsIfNeeded().withMode(CreateMode.EPHEMERAL)
-                            .forPath(instanceUrl);
-                }
+                EXECUTOR.execute(new Runnable() {
+                    @Override
+                    public void run() {
+                        try {
+                            if (null == client.checkExists().forPath(instanceUrl)) {
+                                client.create().creatingParentsIfNeeded().withMode(CreateMode.EPHEMERAL)
+                                        .forPath(instanceUrl);
+                            } else {
+                                client.delete().quietly().forPath(instanceUrl);
+                                client.create().creatingParentsIfNeeded().withMode(CreateMode.EPHEMERAL)
+                                        .forPath(instanceUrl);
+                            }
+                        } catch (Exception e) {
+                            log.error("insert task from nacos to zk was failed, taskId:{}", taskDO.getTaskId(), e);
+                        }
+                    }
+                });
+
                 waitingToAddInstance.add(instanceUrl);
             }
         }
@@ -197,11 +182,9 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
         metaData.put(SkyWalkerConstants.DEST_CLUSTERID_KEY, taskDO.getDestClusterId());
         metaData.put(SkyWalkerConstants.SYNC_SOURCE_KEY,
                 skyWalkerCacheServices.getClusterType(taskDO.getSourceClusterId()).getCode());
-        metaData.put(SkyWalkerConstants.SOURCE_CLUSTERID_KEY, taskDO.getSourceClusterId());
-
+        metaData.put(SOURCE_CLUSTERID_KEY, taskDO.getSourceClusterId());
         String servicePath = monitorPath.computeIfAbsent(taskDO.getTaskId(),
                 key -> convertDubboProvidersPath(metaData.get(DubboConstants.INTERFACE_KEY)));
-
         return convertDubboFullPathForZk(metaData, servicePath, instance.getIp(), instance.getPort());
     }
 
@@ -227,66 +210,14 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
 
     }
 
-
-    private class SyncThread implements Runnable {
-
-        NamingService sourceNamingService;
-
-        TaskDO taskDO;
-
-        CuratorFramework client;
-
-        SyncThread(NamingService sourceNamingService, TaskDO taskDO, CuratorFramework client) {
-            this.sourceNamingService = sourceNamingService;
-            this.taskDO = taskDO;
-            this.client = client;
-        }
-
-        @Override
-        public void run() {
-            try {
-
-                //List<Instance> sourceInstances = sourceNamingService.getAllInstances(taskDO.getServiceName());
-                List<Instance> sourceInstances = sourceNamingService.getAllInstances(taskDO.getServiceName(), taskDO.getGroupName());//fix  with no group
-                Set<String> newInstanceUrlSet = getWaitingToAddInstance(taskDO, client, sourceInstances);
-
-                // 获取之前的备份 删除无效实例
-                deleteInvalidInstances(taskDO, client, newInstanceUrlSet);
-                // 替换当前备份为最新备份
-                instanceBackupMap.put(taskDO.getTaskId(), newInstanceUrlSet);
-                // 尝试恢复因为zk客户端意外断开导致的实例数据
-                tryToCompensate(taskDO, sourceNamingService, filterNeedSync(sourceInstances));
-            } catch (Exception e) {
-                log.error("event process fail, taskId:{}", taskDO.getTaskId(), e);
-                metricsManager.recordError(MetricsStatisticsType.SYNC_ERROR);
-
-            } finally {
-                //serviceNameSet.remove(((NamingEvent) event).getServiceName());//如果考虑高实时性 可以手动remove 这样时间窗口的大小就不固定 依赖处理速度 窗口大小作为兜底
-            }
-        }
-    }
-
-    private List<Instance> filterNeedSync(List<Instance> sourceInstances) {
-        Iterator<Instance> iterator = sourceInstances.iterator();
-        while (iterator.hasNext()) {
-            if (!needSync(iterator.next().getMetadata())) {
-                iterator.remove();
-            }
-        }
-        return sourceInstances;
-    }
-
-    public boolean addSyncService(TaskDO taskDO) {
+    public boolean addSynService(TaskDO taskDO) {
         try {
             NamingService sourceNamingService =
-                    //nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getGroupName());
-                    nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getNameSpace());//fix with no nameSpaceName
+                    nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getNameSpace());
             CuratorFramework client = zookeeperServerHolder.get(taskDO.getDestClusterId(), taskDO.getGroupName());
             nacosListenerMap.putIfAbsent(taskDO.getTaskId(), event -> {
                 if (event instanceof NamingEvent) {
-                    if (serviceNameSet.set(((NamingEvent) event).getServiceName())) {// add event merge
-                        EXECUTOR.execute(new SyncThread(sourceNamingService, taskDO, client));
-                    }
+                    sync(taskDO, client, ((NamingEvent) event).getInstances(), ((NamingEvent) event).getServiceName());
                 }
             });
             sourceNamingService.subscribe(taskDO.getServiceName(), nacosListenerMap.get(taskDO.getTaskId()));
@@ -298,4 +229,31 @@ public class NacosSyncToZookeeperServiceImpl implements SyncService {
         return true;
     }
 
+    public void removeSyncServices(TaskDO taskDO) {
+        try {
+            NamingService sourceNamingService =
+                    nacosServerHolder.get(taskDO.getSourceClusterId(), taskDO.getNameSpace());//fix with no nameSpaceName
+            EventListener eventListener = nacosListenerMap.remove(taskDO.getTaskId());
+            sourceNamingService.unsubscribe(taskDO.getServiceName(), eventListener);
+            sharding.stop(taskDO);
+        } catch (Exception e) {
+            log.error("delete task from nacos to zk was failed, taskId:{}", taskDO.getTaskId(), e);
+            metricsManager.recordError(MetricsStatisticsType.DELETE_ERROR);
+        }
+    }
+
+    private void sync(TaskDO taskDO, CuratorFramework client, List<Instance> sourceInstances, String serviceName) {
+        try {
+            Set<String> newInstanceUrlSet = getWaitingToAddInstance(taskDO, client, sourceInstances);
+            // 获取之前的备份 删除无效实例
+            deleteInvalidInstances(taskDO, client, newInstanceUrlSet);
+            // 替换当前备份为最新备份
+            instanceBackupMap.put(taskDO.getTaskId(), newInstanceUrlSet);
+        } catch (Exception e) {
+            log.error("event process fail, taskId:{}", taskDO.getTaskId(), e);
+            metricsManager.recordError(MetricsStatisticsType.SYNC_ERROR);
+        } finally {
+            //serviceNameSet.remove(taskDO.getServiceName());//如果考虑高实时性 可以手动remove 这样时间窗口的大小就不固定 依赖处理速度 窗口大小作为兜底
+        }
+    }
 }
